resurrect 64bit compile
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2229 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ flasher.cpp
@@ -399,7 +399,7 @@ void Flasher::RenderSetup(RenderDevice* pd3dDevice)
    std::vector<RenderVertex> vvertex;
    GetRgVertex(vvertex);
 
-   numVertices = vvertex.size();
+   numVertices = (int)vvertex.size();
 
    VectorVoid vpoly;
